% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFactorInterations.R
\name{addFactorInteractions}
\alias{addFactorInteractions}
\title{Add a combination of two or more factor variables to a protData object}
\usage{
addFactorInteractions(protdata, basecols, name = NULL)
}
\arguments{
\item{protdata}{A \code{\link[=protdata-class]{protdata}} object of which the data slot needs to be expanded.}

\item{basecols}{The name of the columns in the existing data on which you base your new column.}

\item{name}{The name of the new variable. If no name is provided (default: \code{name=NULL}), the names of the \code{basecols} will be concatenated separated by an "_".}
}
\value{
A \code{\link[=protdata-class]{protdata}} object with an added column in its data slot containing the new variable.
}
\description{
Adds a categorical variable to the data slot of a \code{\link[=protdata-class]{protdata}} that is a combination of two or more exisiting factor variables.
}
\examples{
#.........
}
