% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{adjustNames}
\alias{adjustNames}
\title{Adjust the names of the elements in a list of dataframes}
\usage{
adjustNames(datalist, predictors)
}
\arguments{
\item{datalist}{A list of dataframes.}

\item{predictors}{A vector of predictors corresponding to some of the columnames in the list of dataframes for which the given adjustment will be performed if the columns corresponding to these predictors are characters or factors.}
}
\value{
The list of dataframes with adusted elements.
}
\description{
Given a list of dataframes, this function pastes the colnames of all factors and characters to the elements in these columns.
This is necessary to produces consisten results when fitting with lm or lmer. This function is borderline internal.
}
