% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPeptides.R
\name{countPeptides}
\alias{countPeptides}
\title{Count the number of peptides in each group}
\usage{
countPeptides(protdata, pData, predictor)
}
\arguments{
\item{protdata}{A \code{\link[=protdata-class]{protdata}} object}

\item{pData}{A dataframe that contains experimenter-supplied variables describing phenotypes and/or experimental factors that might influence the quantitative value of interest for each mass spec run. The \code{colnames} should be all the different effects that influence the outcome. The first column should hold all the different run names. The corresponding levels of the other effects should be in the subsequent columns.}

\item{predictor}{A predictor for which you would like to summarize the number of identified peptides per level for each protein. This predictor should correspond to one element of the \code{colnames} of \code{pData}.}
}
\description{
Given a \code{\link[=protdata-class]{protdata}} object, a dataframe pData containing experiment information and a predictor of interest,
returns a data frame giving the amount of peptides observed for each level of the predictor.
}
