% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_p_adjust.R
\name{cutOffPval}
\alias{cutOffPval}
\title{Find cut off for p-values enriched in values close to 1}
\usage{
cutOffPval(p)
}
\arguments{
\item{p}{A vector of p-values.}
}
\value{
A named vector with two values: "cutoff_l", the maximal p-value below the cut-off and "cutoff_u", the minimal p-value that is above the cut-off.
}
\description{
Contrast tests based on mixed models often produce p-values containing an excess of values close to 1.
This function will, given a vector of p-values find a cut-off above which the values are excessively enriched.
The cut-off is determined based on the difference coefficient and 2nd order difference coefficient of sequential p-values.
Therefore, it will (almost) always find a cut-off, even if there is no real enrichment in p-values.
One should thus always assess whether there is truly an enrichment when using this function.
}
\examples{
.......
}
