% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrtool_subset.R
\name{fdrtool_subset}
\alias{fdrtool_subset}
\title{Estimate (Local) False Discovery Rates For Diverse Test Statistics}
\usage{
fdrtool_subset(x, x_red = x, statistic = c("normal", "correlation",
  "pvalue"), verbose = TRUE, cutoff.method = c("fndr", "pct0", "locfdr"),
  pct0 = 0.75)
}
\arguments{
\item{x}{A vector of the observed test statistics.}

\item{x_red}{A subset vector of x that will be given to the fdrtool algorithm.}

\item{statistic}{One of "normal" (default), "correlation", "pvalue". This species the null model.}

\item{verbose}{A logical indicating if status messages should be printed out. Defaults to \code{TRUE}.}

\item{cutoff.method}{One of "fndr" (default), "pct0", "locfdr".}

\item{pct0}{The fraction of data x_red that will be used for fitting null model - only used if cutoff.method="pct0".}
}
\value{
A list with the following components:
  \itemize{
  \item \code{x} The input vector of the observed test statistics.
  \item \code{x_red} The subset of the input vector of the observed test statistics that was given to the fdrtool algorithm.
  \item \code{pval}	A vector with adjusted p-values for each element of \code{x_red}.
  \item \code{pval_full} A vector with adjusted p-values for each element of \code{x}.
  \item \code{qval}	A vector with q-values (Fdr) for each element of \code{x_red}.
  \item \code{qval_full} A vector with q-values (Fdr) for each element of \code{x}.
  \item \code{lfdr}	A vector with local fdr values for each element of \code{x_red}.
  \item \code{lfdr_full} A vector with local fdr values for each element of \code{x}.
  \item \code{statistic} The specified type of null model.
  \item \code{param} A vector containing the estimated parameters (\code{eta0} (the null proportion of \code{x_red}) and the free parameter of the null model).
  \item \code{x0} The cut off value of \code{x_red}, indicating the separation between the null model and the mixture model.
  \item \code{f.pval}, \code{F.pval}, \code{fdr.pval}, \code{Fdr.pval}, \code{f0}, \code{F0}, \code{get.pval}, \code{fdr}, \code{Fdr} Functions needed to make diagnostic plots.
}
}
\description{
This function is largely based on the \code{\link[fdrtool]{fdrtool}} function from the \pkg{fdrtool} package by Korbinian Strimmer (\url{http://strimmerlab.org}).
\pkg{fdrtool} takes a vector of z-scores (or of correlations, p-values, or t-statistics), and estimates for each case both the tail area-based Fdr as well as the density-based fdr (=q-value resp. local false discovery rate). The parameters of the null distribution are estimated adaptively from the data (except for the case of p-values where this is not necessary).
Our only modification is adding the possibility to provide a reduced input (e.g. an input of which the excess in p values equal to 1 are removed) for estimating the null distribution while still correcting for the total number of input values.
}
\details{
See package \pkg{fdrtool}.
}
\examples{
.......
}
\references{
Strimmer, K. (2008a). A unified approach to false discovery rate estimation. BMC Bioinformatics 9: 303. Available from \url{http://www.biomedcentral.com/1471-2105/9/303/}.

Strimmer, K. (2008b). fdrtool: a versatile R package for estimating local and tail area- based false discovery rates. Bioinformatics 24: 1461-1462. Available from \url{http://bioinformatics.oxfordjournals.org/cgi/content/abstract/24/12/1461}.
}
