% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{getAccessions,protdata-method}
\alias{getAccessions,protdata-method}
\title{Extract the accession slot of a protdata object}
\usage{
\S4method{getAccessions}{protdata}(object)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{accession} slot needs to be extracted.}
}
\value{
A factor vector containing the accessions present in the \code{\link[=protdata-class]{protdata}} object.
}
\description{
\code{getAccessions} extracts the \code{accession} slot from a \code{\link[=protdata-class]{protdata}} object.
}

