% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{methods}
\name{getAccessions,protLM-method}
\alias{getAccessions,protLM-method}
\title{Extract the accession slot of a protLM object}
\usage{
\S4method{getAccessions}{protLM}(object)
}
\arguments{
\item{object}{A \code{\link[=protLM-class]{protLM}} object of which the \code{accession} slot needs to be extracted.}
}
\value{
A factor vector containing the accessions present in the \code{\link[=protLM-class]{protLM}} object.
}
\description{
\code{getAccessions} extracts the \code{accession} slot from a \code{\link[=protLM-class]{protLM}} object.
}
