% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{getAnnotations,protdata-method}
\alias{getAnnotations,protdata-method}
\title{Extract the annotation slot of a protdata object}
\usage{
\S4method{getAnnotations}{protdata}(object)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{annotation} slot needs to be extracted.}
}
\value{
A character matrix containing the annotations present in the \code{\link[=protdata-class]{protdata}} object. Rows correspond to accessions, while columns correspond to different annotations.
}
\description{
\code{getAnnotations} extracts the \code{annotation} slot from a \code{\link[=protdata-class]{protdata}} object.
}

