% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBetaVcovDf.R
\docType{methods}
\name{getBetaVcovDf,lm-method}
\alias{getBetaVcovDf,lm-method}
\title{Get beta, vcov, df and sigma from a general linear model}
\usage{
\S4method{getBetaVcovDf}{lm}(model, exp_unit = NULL, pars_df = NULL,
  subset = NULL)
}
\arguments{
\item{model}{A general linear model object of class \code{\link[=lm-class]{lm}}.}

\item{exp_unit}{The effect in the model that corresponds to the experimental unit. Only needed when one would like to calculate a more conservative way of estimating the degrees of freedom.
The default way (\code{exp_unit=NULL}) estimates the degrees of freedom by substracting the total number of observations by the number of parameters. However, often, observations are not completely independent. A more conservative way (\code{df_exp}) is defining on which level the treatments were executed and substracting all degrees of freedom lost due to between-treatement effects (\code{pars_df}) from the number of treatments.}

\item{pars_df}{Only used if \code{exp_unit} is not \code{NULL}. Character vector indicating all parameters in the model that are between-treatment effects in order to calculate a more conservative degrees of freedom (\code{df_exp}). If left to default (\code{NULL}), all parameters in the model will be asumed to be between-treatment effects (this is not adviced as the result will mostly be too conservative).}

\item{subset}{...... !!!!Experimental feature not thoroughly tested!!!!}
}
\value{
A list containing (1) a named column matrix beta containing the parameter estimates, (2) a named square variance-covariance matrix, (3) a numeric value equal to the residual degrees of freedom, (4) a numeric value equal to the residual standard deviation of the model and (5) \code{NULL} if \code{exp_unit} is left to its default value \code{NULL}, else: a conservative estimate of the degrees of freedom based on the number of experimental units and the degrees of freedom lost due to between-treatment effects.
}
\description{
This function returns a list containing the parameter estimates \code{beta}, the variance-covariance matrix \code{vcov},
the residual degrees of freedom \code{df} and the residual standard deviation \code{sigma} based on a general linear model fitted by the \code{lm} function from the stats package.
This function will only rarely be called by the end-user. When calculating these values for \code{\link[=protLM-class]{protLM}} objects, we recommend using the function \code{getbetaVcovDfList}.
}
\examples{
data(proteinsCPTAC, package="MSqRob")
lmmodel <- lm(formula="value ~ 1 + conc + instrlab + Sequence",data=getData(proteinsCPTAC[2]))
getBetaVcovDf(lmmodel)
}
