% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBetaVcovDf.R
\docType{methods}
\name{getBetaVcovDf,lmerMod-method}
\alias{getBetaVcovDf,lmerMod-method}
\title{Get beta, vcov, df and sigma from a mixed model}
\usage{
\S4method{getBetaVcovDf}{lmerMod}(model, exp_unit = NULL,
  pars_between = NULL, Ginvoffset = 1e-18)
}
\arguments{
\item{model}{A linear mixed effects model object of class \code{\link[=lmerMod-class]{lmerMod}}.}

\item{exp_unit}{The effect in the model that corresponds to the experimental unit. Only needed when one would like to calculate a more conservative way of estimating the degrees of freedom.
The default way of estimating the degrees of freedom (\code{exp_unit=NULL}) subtracts the total number of observations by the trace of the Hat matrix. However, often, observations are not completely independent. A more conservative way (\code{df_exp}) is defining on which level the treatments were executed and substracting all degrees of freedom lost due to between-treatement effects (\code{pars_between}) from the number of treatments.}

\item{pars_between}{Only used if exp_unit is not \code{NULL}. Character vector indicating all parameters in the model that are between-treatment effects in order to calculate a more conservative degrees of freedom (\code{df_exp}). If left to default (\code{NULL}), all parameters in the model will be asumed to be between-treatment effects (this is not adviced as the result will mostly be too conservative).}

\item{Ginvoffset}{A numeric value indicating the offset added the the diagonal of the G matrix to prevent near-singularity. Defaults to \code{1e-18}.}
}
\value{
A list containing (1) a named column matrix beta containing the parameter estimates, (2) a named square variance-covariance matrix, (3) a numeric value equal to the residual degrees of freedom and (4) a numeric value equal to the residual standard deviation of the model.
}
\description{
This function returns a list containing the parameter estimates \code{beta}, the variance-covariance matrix \code{vcov},
the residual degrees of freedom \code{df} and the residual standard deviation \code{sigma} based on a mixed effects model fitted by the \code{lmer} function from the lme4 package.
This function will only rarely be called by the end-user. When calculating these values for \code{\link[=protLM-class]{protLM}} objects, we recommend using the function \code{\link{getBetaVcovDfList}}.
The variance covariance matrix is bias-adjusted, the degrees of freedom are calculated as the trace of the Hat matrix (Ruppert et al., 2003).
}
\examples{
data(proteinsCPTAC, package="MSqRob")
mixedmodel <- lmer(formula="value ~ 1 + (1|conc) + (1|instrlab) + (1|Sequence)",data=getData(proteinsCPTAC[2]))
getBetaVcovDf(mixedmodel)
}
\references{
David Ruppert, M.P. Want and R.J. Carroll.
Semiparametric Regression.
 Cambridge University Press, 2003.
}
