% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{getData,protdata-method}
\alias{getData,protdata-method}
\title{Extract the data slot of a protdata object}
\usage{
\S4method{getData}{protdata}(object, simplify = TRUE)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{data} slot needs to be extracted.}

\item{simplify}{Logical. Should the list of data frames be simplified to a data frame if the length of this list equals 1?}
}
\value{
The list of data frames corresponding to the data present in the \code{\link[=protdata-class]{protdata}} object.
}
\description{
\code{getData} extracts the \code{data} slot from a \code{\link[=protdata-class]{protdata}} object.
}
