% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{methods}
\name{getModels,protLM-method}
\alias{getModels,protLM-method}
\title{Extract the model slot of a protLM object}
\usage{
\S4method{getModels}{protLM}(object, simplify = TRUE)
}
\arguments{
\item{object}{A \code{\link[=protLM-class]{protLM}} object of which the \code{model} slot needs to be extracted.}

\item{simplify}{Logical. Should the list of models be simplified to a model frame if the length of this list equals 1?}
}
\value{
The list of models corresponding to the accessions present in the \code{\link[=protLM-class]{protLM}} object.
}
\description{
\code{getModels} extracts the \code{model} slot from a \code{\link[=protLM-class]{protLM}} object.
}

