% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{getPData,protdata-method}
\alias{getPData,protdata-method}
\title{Extract the annotation slot of a protdata object}
\usage{
\S4method{getPData}{protdata}(object)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{pData} slot needs to be extracted.}
}
\value{
A data frame containing the experimental annotation of the \code{\link[=protdata-class]{protdata}} object.
}
\description{
\code{getPData} extracts the \code{pData} slot from a \code{\link[=protdata-class]{protdata}} object.
}
