% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeezePars.R
\name{getThetaVars}
\alias{getThetaVars}
\title{Get variances and degrees of freedom of model parameters}
\usage{
getThetaVars(protLM, par_names)
}
\arguments{
\item{protLM}{A \code{\link[=protLM-class]{protLM}} object of which residual variances and/or model parameters need to be returned.}

\item{par_names}{Character vector indicating of which model parameters the variance needs to be returned. When squeezing random effects, provide their names. Fixed effects are present in shrinkage groups, e.g. ridgeGroup.1. If you want to return their variance as well, provide the names of the shrinkage groups that need to be squeezed. If \code{par_names=NULL}, \code{NA} will be returned in the \code{thetas} and \code{df_thetas} slots of the output.}
}
\value{
A named list with 4 slots. The first slot \code{thetas} contains a matrix with in each column the estimated variances for an effect specified in the \code{par_names} argument, each row corresponds to a different accession in the \code{\link[=protLM-class]{protLM}} object.
The second slot \code{df_thetas} contains a matrix of similar structure to \code{thetas} but containing the degrees of freedom corresponding to the estimated variances. The third slot \code{vars} contains a vector of residual variances for each accession and the fourth slot \code{df_vars} contains a vector of residual degrees of freedom.
}
\description{
This function extracts the estimated variances of specified random effects and shrunken fixed effects as well as their associated degrees of freedom (based on the trace of the Hat matrix). It also always returns the residual variance and the residual degrees of freedom.
}
\examples{
....
}
\references{
....
}

