% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\name{index2Numeric}
\alias{index2Numeric}
\title{Convert index to only numeric values}
\usage{
index2Numeric(i, accessions)
}
\arguments{
\item{i}{A vector of indices (either numeric or a character or factor vector corresponding to elements in the \code{accessions} vector).}

\item{accessions}{A character or factor vector containing accessions.}
}
\value{
A numeric vector containing either the original numeric values or the numeric values corresponding to the position of each element of \code{i} in the \code{accessions} vector.
}
\description{
This function converts a vector of indices, which are a potential mixture of numeric values and accessions (characters or even factors)
to a vector which retains the numeric indices, but converts accessions to their corresponding index in a provided \code{accessions} vector.
}

