% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_hlpFunctions.R
\name{init_ann_MQ_Excel}
\alias{init_ann_MQ_Excel}
\title{Initialize an annotation Excel file based on a MaxQuant peptides.txt file}
\usage{
init_ann_MQ_Excel(file, savepath = NULL,
  output_name = "experimental_annotation", col_name = "run",
  pattern = "Intensity.", remove_pattern = TRUE)
}
\arguments{
\item{file}{The name of a MaxQuant peptides.txt file. For more details about how this argument can be specified, see \code{\link[utils]{read.table}}.}

\item{savepath}{The file path to the directory where you want the output Excel annotation file to be saved. If set to \code{NULL} (the default), the Excel file will be saved in the working directory.}

\item{output_name}{The name of the output Excel annotation file. Defaults to \code{"experimental_annotation"}.}

\item{col_name}{The name of the only column created in the Excel file. Defaults to \code{"run"}.}

\item{pattern}{A character string containing a regular expression that will be matched to the file's header. The columns matching the expression should be the columns containing the peptide intensity values. Defaults to "Intensity.".}

\item{remove_pattern}{A logical indicating whether the expression in "pattern" should be removed from the column names in the resulting Excel file. Defaults to \code{TRUE}.}
}
\value{
An Excel file with one column containing the different run names present in the MaxQuant peptides.txt file and column name \code{col_name}.
}
\description{
Creates an Excel file with one column containing the run names of an experiment based on the column names of a MaxQuant peptides.txt file. This function might be a start in constructing the first column of the annatation Excel file, other colunms need to be added manually.
}

