% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saves_loads_MSqRob.R
\name{inspect_loads_MSqRob}
\alias{inspect_loads_MSqRob}
\title{Inspect the variables present in a data.frame from binary file}
\usage{
inspect_loads_MSqRob(file = NULL)
}
\arguments{
\item{file}{character string: the (RDatas) filename from which to inspect the variable names.}
}
\value{
The variable names present in the .RDatas file.
}
\description{
\code{inspect_loads_MSqRob} returns the names of the variables present in a .RDatas file, some or all of which can be used in the \code{variables} argument of the \code{\link{loads_MSqRob}} function if one prefers to load only some variables.
}

