% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{length,protdata-method}
\alias{length,protdata-method}
\title{Length of a protdata object}
\usage{
\S4method{length}{protdata}(x)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the number of elements needs to be determined.}
}
\value{
An integer value indicating the number of elements in a \code{\link[=protdata-class]{protdata}} object.
}
\description{
The \code{length} method counts the number of elements in a \code{\link[=protdata-class]{protdata}} object.
}
