% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{methods}
\name{length,protLM-method}
\alias{length,protLM-method}
\title{Length of a protLM object}
\usage{
\S4method{length}{protLM}(x)
}
\arguments{
\item{object}{A \code{\link[=protLM-class]{protLM}} object of which the number of elements needs to be determined.}
}
\value{
An integer value indicating the number of elements in a \code{\link[=protLM-class]{protLM}} object.
}
\description{
The \code{length} method counts the number of elements in a \code{\link[=protLM-class]{protLM}} object.
}
