% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnSet2protdata.R
\name{MSnSet2protdata}
\alias{MSnSet2protdata}
\title{MSnSet to protein data}
\usage{
MSnSet2protdata(MSnSet, accession = NULL, annotations = NULL,
  quant_name = "quant_value", printProgress = FALSE, shiny = FALSE,
  message = NULL)
}
\arguments{
\item{MSnSet}{An \code{\link[=MSnSet-class]{MSnSet}} object that needs to be converted to a \code{\link[=protdata-class]{protdata}} object.}

\item{accession}{A character string or numeric index indicating the column in the fData slot of object  \code{\link[=MSnSet-class]{MSnSet}} that contains the identifiers by which the data should be grouped in the resulting \code{\link[=protdata-class]{protdata}} object (typically the protein identifier).}

\item{annotations}{A vector of character strings or numeric indices indicating the columns in the fData slot of object  \code{\link[=MSnSet-class]{MSnSet}} that contain additional information on the accessions (typically protein names, gene names, gene ontologies,...) that should be retained. In case multiple values of the same annotation column would exist for a unique accession, these values are pasted together. Defaults to \code{NULL}, in which case no annotations will be added.}

\item{quant_name}{A character string indicating the name to be given to the column that will contain the quantitative values of interest (mostly peptide intensities or peptide areas under the curve). Defaults to \code{"quant_value"}.}

\item{printProgress}{A logical indicating whether the R should print a message before converting each accession. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user, or \code{NULL} to hide the current message (if any).}
}
\value{
A \code{\link[=protdata-class]{protdata}} object.
}
\description{
Converts an \code{\link[=MSnSet-class]{MSnSet}} object to a a \code{\link[=protdata-class]{protdata}} object, which can be used for further analysis.
}
\examples{
#This example will convert MSnSet object peptides into a protdata object proteins.
library(MSnbase)
colInt <- grepEcols(system.file("extdata/CPTAC", "peptides.txt", package = "MSqRob"), pattern="Intensity.", split = "\\t")
#Import the data as an MSnSet object
peptides <- readMSnSet2(system.file("extdata/CPTAC", "peptides.txt", package = "MSqRob"), ecol = colInt, sep = "\\t")
fData(peptides) <- fData(peptides)[,c("Proteins","Sequence","PEP")]
#Do this only for the first 50 peptide to save execution time
proteins <- MSnSet2protdata(peptides[1:50], "Proteins")
}
