% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_hlpFunctions.R
\name{pasteAnnotation}
\alias{pasteAnnotation}
\title{Add annotations to a data frame in "long" format}
\usage{
pasteAnnotation(df, run_col, run_col_annot, pData)
}
\arguments{
\item{df}{A data frame in "long" format to which extra annotation columns need to be added.}

\item{run_col}{A character indicating the column in the data frame \code{df} that contains a different identifier for each mass spec run. All elements in this column should be present exactly once in exactly one column of the experimental annotation data frame \code{pData} and be equal to the elements of the \code{run_col_annot} column in the \code{pData} data frame.}

\item{run_col_annot}{A character indicating the column in the annotation data frame \code{pData} that contains a different identifier for each mass spec run. All elements in this column should be present exactly once in this column of the experimental annotation data frame \code{pData} and be equal to the elements of the \code{run_col} column in the data frame \code{df}.}

\item{pData}{A data frame containing the experiment annotation. Exactly one colum in the experiment annotation should contain the mass spec run names.}
}
\value{
A data frame to which the annotations are left-joined.
}
\description{
This function adds extra columns to a data frame in "long" format (i.e. the data frame has one observation row per measurement (thus, multiple rows per subject)) based on either a path to an annotation file or an annotation data frame.
}
