% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrtool_subset.R
\name{plot_fdrtool}
\alias{plot_fdrtool}
\title{Make diagnostic plots for fdrtool results}
\usage{
plot_fdrtool(result, make_plots = c(TRUE, TRUE, TRUE), color.figure = TRUE,
  verbose = TRUE)
}
\arguments{
\item{result}{The result of a call to the \code{\link{fdrtool_subset}} function.}

\item{make_plots}{A vector of three logical values. Each value indicates whether a certain plot should be made. Defaults to \code{c(TRUE, TRUE, TRUE)} indicating that all three plots should be made.}

\item{color.figure}{A logical indicating whether a color figure or a black and white figure is produced. Defaults to \code{TRUE}, i.e. to color figure.}

\item{verbose}{A logical indicating if status messages should be printed out. Defaults to \code{TRUE}.}
}
\value{
Diagnostic plots.
}
\description{
This function allows to make the three diagnostic plots from the \pkg{fdrtool} package by Korbinian Strimmer (\url{http://strimmerlab.org}) in separate windows.
The implementation is identical to that in the \pkg{fdrtool} package, except that it allows to plot the functions separately.
The first plot is also different in showing a histogram for both the vector of the observed test statistics \code{x} in white
and a histogram for the subset of this vector that was used in the \code{\link{fdrtool_subset}} function in gray.
}
\examples{
.......
}
\references{
Strimmer, K. (2008a). A unified approach to false discovery rate estimation. BMC Bioinformatics 9: 303. Available from \url{http://www.biomedcentral.com/1471-2105/9/303/}.

Strimmer, K. (2008b). fdrtool: a versatile R package for estimating local and tail area- based false discovery rates. Bioinformatics 24: 1461-1462. Available from \url{http://bioinformatics.oxfordjournals.org/cgi/content/abstract/24/12/1461}.
}
