% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_wide.R
\name{preprocess_wide}
\alias{preprocess_wide}
\title{Preprocess data in "wide" format}
\usage{
preprocess_wide(df, accession, split, exp_annotation = NULL,
  type_annot = NULL, quant_cols, aggr_by, aggr_function = "sum",
  logtransform = TRUE, base = 2, normalisation = "quantiles",
  smallestUniqueGroups = TRUE, useful_properties = NULL, filter = NULL,
  filter_symbol = NULL, minIdentified = 2, colClasses = NA,
  printProgress = FALSE, shiny = FALSE, ...)
}
\arguments{
\item{df}{A data frame that contains data in "wide" format.}

\item{accession}{A character indicating the column that contains the unit on which you want to do inference (typically the protein identifiers).}

\item{split}{A character indicating which string is used to separate accession groups.}

\item{exp_annotation}{Either the path to the file which contains the experiment annotation or a data frame containing the experiment annotation. Exactly one colum in the experiment annotation should contain the mass spec run names. Annotation in a file can be both a tab-delimited text document or an Excel file. For more details, see \code{\link[utils]{read.table}} and \code{\link[openxlsx]{read.xlsx}}. As an error protection measurement, leading and trailing spaces in each column are trimmed off. The default, \code{NULL} indicates there is no annotation to be added.}

\item{type_annot}{If \code{exp_annotation} is a path to a file, the type of file. \code{type_annot} is mostly obsolete as supported files will be automatically recognized by their extension. Currently only \code{"tab-delim"} (tab-delimited file), \code{"xlsx"} (Office Open XML Spreadsheet file) and \code{NULL} (file type decided based on the extension) are supported. If the extension is not recognized, the file will be assumed to be a tab-delimited file. Defaults to \code{NULL}.}

\item{quant_cols}{Either a character or numeric vector indicating the columns that contain the quantitative values of interest (mostly peptide intensities or peptide areas under the curve) or a character string of length one indicating a pattern that is unique for the column names of the columns that contain the quantitative values of interest.}

\item{aggr_by}{A character indicating the column by which the data should be aggregated. We advise to aggregate the data by peptide sequence (thus aggregate over different charge states and modification statuses of the same peptide). If you only want to aggregate over charge states, set \code{aggr_by} to the column corresponding to the modified sequences. If no aggregation at all is desired, create a new column in which you paste together the modified sequences and the charge states.}

\item{aggr_function}{The function used to aggregate intensity data. Defaults to \code{"sum"}.}

\item{logtransform}{A logical value indicating whether the intensities should be log-transformed. Defaults to \code{TRUE}.}

\item{base}{A positive or complex number: the base with respect to which logarithms are computed. Defaults to 2.}

\item{normalisation}{A character vector of length one that describes how to normalise the data frame \code{df}. See \code{\link[=normalise-methods]{normalise}} for details. Defaults to \code{"quantiles"}. If no normalisation is wanted, set \code{normalisation="none"}.}

\item{smallestUniqueGroups}{A logical indicating whether proteins groups for which any of its member proteins is present in a smaller protein group should be removed from the dataset. Defaults to \code{TRUE}.}

\item{useful_properties}{The columns of the data frame \code{df} that are useful in the further analysis and/or inspection of the data and should be retained. Defaults to \code{NULL}, indicating no extra columns should be retained.}

\item{filter}{A vector of names corresponding to the columns in the data frame \code{df} that contain a \code{filtersymbol} that indicates which rows should be removed from the data.
Typical examples are contaminants or reversed sequences. Defaults to \code{NULL}, indicating no filtering should be applied.}

\item{filter_symbol}{A character indicating the symbol in the columns corresponding to the \code{filter} argument that is used to indicate rows that should be removed from the data. Defaults to \code{NULL}.}

\item{minIdentified}{A numeric value indicating the minimal number of times a peptide sequence should be identified in the dataset in order not to be removed. Defaults to 2.}

\item{colClasses}{character. Only used when the \code{exp_annotation} argument is a filepath. A vector of classes to be assumed for the columns of the experimental annotation data frame. Recycled if necessary. If named and shorter than required, names are matched to the column names with unspecified values are taken to be NA.
Possible values are \code{NA} (the default, when \code{type.convert} is used), \code{NULL} (when the column is skipped), one of the atomic vector classes (\code{logical}, \code{integer}, \code{numeric}, \code{complex}, \code{character}, \code{raw}), or \code{factor}, \code{Date} or \code{POSIXct}. Otherwise there needs to be an as method (from package \code{methods}) for conversion from \code{character} to the specified formal class.}

\item{printProgress}{A logical indicating whether the R should print a message before performing each preprocessing step. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{...}{Extra parameters to be passed to the normalisation functions.}
}
\value{
A preprocessed \code{\link[=MSnSet-class]{MSnSet}} object that is ready to be converted into a \code{\link[=protdata-class]{protdata}} object.
}
\description{
Performs a standard preprocessing pipeline on data frames in "wide" format (i.e. the data frame has one observation row per subject with each measurement present as a different variable).
By default, intensity values are log2 transformed and then quantile normalized. Next, the \code{\link[=smallestUniqueGroups]{smallestUniqueGroups}} function is applied,
which removes proteins groups for which any of its member proteins is present in a smaller protein group. Then, unwanted sequences (such as reverse sequences or unwanted sequences) are filtered out.
Next, irrelevant columns are dropped. Then, peptide sequences that are identified only once in a single mass spec run are removed because with only 1 identification, the model will be perfectly confounded. Finally, potential experimental annotations are added to the data frame.
}

