% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_p_adjust_protwise.R
\name{prot.p.adjust_protwise}
\alias{prot.p.adjust_protwise}
\title{Adjust P-values for Multiple Comparisons per Protein}
\usage{
prot.p.adjust_protwise(coefmatlist, L, method = "Schaffer", pcol = "pval",
  qname = "qval", stage2 = FALSE, significant_stage1 = NULL)
}
\arguments{
\item{coefmatlist}{Either a matrix containing a column with p-values or a list of such matrices.}

\item{L}{A contrast matrix with the parameter levels as rows and a column for each contrast.}

\item{method}{Correction method. Can be abbreviated. Defaults to "Holm", which is currently the only available option.}

\item{pcol}{A character string or numeric index indicating the column containing the p-values. Defaults to "pval".}

\item{qname}{A character string indicating the name that will be given to the column containing the q-values. Defaults to "qval".}

\item{stage2}{Set this to \code{TRUE} when using this function in the 2nd stage of a stage-wise test.}

\item{significant_stage1}{Only used when \code{stage2 = TRUE}. This should be a character vector of protein identifiers that were retained in the first stage during stage-wise testing.}
}
\value{
Depending on the input, either a matrix or a list of matrices with an extra numeric column containing corrected p-values.
}
\description{
Given matrix or a list of matrices with a column containing p-values and a contrast matrix, adds a column containing Holm-adjusted p-values correcting for the number of independent contrasts.
This function can also be used to calculate the corrected p-values in the second stage when performing stage-wise testing.
}
\examples{
#....
}
