% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{class}
\name{protLM-class}
\alias{protLM-class}
\title{Protein Linear Model Object - class}
\description{
An S4 class for storing the model fits of feature- or peptide-level data of a set of LC-MS proteomics experiments, grouped by protein.
Objects of this class are supplied as an argument to the \code{\link{test.protLMcontrast}} function.
}
\section{Slots}{

\describe{
\item{\code{protein}}{A factor vector with length \code{k} equal to the total number of proteins or protein groups in the data.}

\item{\code{model}}{A list of length \code{k} containing the corresponding fitted models.}
}}

