% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R
\name{read_MaxQuant}
\alias{read_MaxQuant}
\title{Import a MaxQuant peptides.txt file}
\usage{
read_MaxQuant(file, pattern = "Intensity.", remove_pattern = TRUE)
}
\arguments{
\item{file}{The name of a MaxQuant peptides.txt file. For more details about how this argument can be specified, see \code{\link[utils]{read.table}}.}

\item{pattern}{A character string containing a regular expression that will be matched to the file's header. The columns matching the expression should be the columns containing the peptide intensity values. Defaults to "Intensity.".}

\item{remove_pattern}{A logical indicating whether the expression in "pattern" should be removed from the column names in the resulting \code{\link[=MSnSet-class]{MSnSet}} object. Defaults to \code{TRUE}.}
}
\value{
An object of class \code{\link[=MSnSet-class]{MSnSet}}.
}
\description{
Imports a MaxQuant peptides.txt file and converts it into an \code{\link[=MSnSet-class]{MSnSet}} object.
}

