% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R
\name{read_moFF}
\alias{read_moFF}
\title{Import a peptides summary file produced by moFF}
\usage{
read_moFF(file, pattern = "sumIntensity_", remove_pattern = TRUE,
  shiny = FALSE, message = NULL)
}
\arguments{
\item{file}{The name of a moFF .tab file. For more details about how this argument can be specified, see \code{\link[utils]{read.table}}.}

\item{pattern}{A character string containing a regular expression that will be matched to the file's header. The columns matching the expression should be the columns containing the peptide intensity values. Defaults to "?????".}

\item{remove_pattern}{A logical indicating whether the expression in "pattern" should be removed from the column names in the resulting \code{\link[=MSnSet-class]{MSnSet}} object. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user, or \code{NULL} to hide the current message (if any).}
}
\value{
An object of class \code{\link[=MSnSet-class]{MSnSet}}.
}
\description{
Imports a moFF (Argentini et al., 2016) .tab file and converts it into an \code{\link[=MSnSet-class]{MSnSet}} object (Gatto et al., 2012).
}
\references{
Gatto L, Lilley KS. MSnbase - an R/Bioconductor package for isobaric tagged mass spectrometry data visualization, processing and quantitation. Bioinformatics. 2012 Jan 15;28(2):288-9. \url{https://doi.org/10.1093/bioinformatics/btr645}.

Argentini A,	Goeminne LJE,	Verheggen K,	Hulstaert N,	Staes A, Clement L	& Martens L. moFF: a robust and automated approach to extract peptide ion intensities. Nature Methods. 2016 13:964–966.  \url{http://www.nature.com/nmeth/journal/v13/n12/full/nmeth.4075.html}.
}
