% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{selectAccessions,protdata-method}
\alias{selectAccessions,protdata-method}
\title{Select accessions from a protdata object}
\usage{
\S4method{selectAccessions}{protdata}(object, i, keep = TRUE)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which certain elements will be selected or removed.}

\item{i}{Either a character or a factor vector containing accessions present in the \code{accessions} slot of the \code{\link[=protdata-class]{protdata}} object or a numeric vector containing indices pointing to certain elements in the the \code{\link[=protdata-class]{protdata}} object or a combination of both (deprecated).}

\item{keep}{A logical indicating whether the selected elements should be retained (\code{keep==TRUE}) or removed (\code{keep==FALSE}). Defaults to \code{TRUE}.}
}
\value{
A reduced \code{\link[=protdata-class]{protdata}} object containing only the elements specified in \code{i} (if \code{keep==TRUE}) or all elements except those specified in \code{i} (if \code{keep==FALSE}).
}
\description{
Select certain elements from a \code{\link[=protdata-class]{protdata}} object based on one or more accession and/or numeric indices. If \code{keep==FALSE}, these indices are omitted instead of selected. Note that for selecting, one could also use the "[" method.
}
