% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{setAccessions,protdata-method}
\alias{setAccessions,protdata-method}
\title{Change the accession slot of a protdata object}
\usage{
\S4method{setAccessions}{protdata}(object, accessions)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{accession} slot needs to be modified.}

\item{accessions}{A factor or character vector of accessions which will be inserted in the \code{accession} slot of the \code{\link[=protdata-class]{protdata}} object. Note that the length of this vector should be equal to the number of elements in the \code{data} slot and the number of rows in the \code{annotation} slot.}
}
\value{
A \code{\link[=protdata-class]{protdata}} object of which the \code{accession} slot is replaced by the \code{accessions} factor vector.
}
\description{
\code{setAccessions} changes the \code{accession} slot from a \code{\link[=protdata-class]{protdata}} object to a prespecified factor vector.
}
