% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{methods}
\name{setAccessions,protLM-method}
\alias{setAccessions,protLM-method}
\title{Change the accession slot of a protLM object}
\usage{
\S4method{setAccessions}{protLM}(object, accessions)
}
\arguments{
\item{object}{A \code{\link[=protLM-class]{protLM}} object of which the \code{accession} slot needs to be modified.}

\item{accessions}{A factor or character vector of accessions which will be inserted in the \code{accession} slot of the \code{\link[=protLM-class]{protLM}} object. Note that the length of this vector should be equal to the number of models in the \code{model} slot and the number of rows in the \code{annotation} slot.}
}
\value{
A \code{\link[=protLM-class]{protLM}} object of which the \code{accession} slot is replaced by the \code{accessions} factor vector.
}
\description{
\code{setAccessions} changes the \code{accession} slot from a \code{\link[=protLM-class]{protLM}} object to a prespecified factor vector.
}
