% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{setAnnotations,protdata-method}
\alias{setAnnotations,protdata-method}
\title{Change the annotation slot of a protdata object}
\usage{
\S4method{setAnnotations}{protdata}(object, annotations)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{annotation} slot needs to be modified.}

\item{annotations}{A matrix which will be inserted in the \code{annotation} slot of the \code{\link[=protdata-class]{protdata}} object. Note that the number of rows of this matrix should be equal to the number of elements in both the \code{accession} and the \code{data} slot.}
}
\value{
A \code{\link[=protdata-class]{protdata}} object of which the \code{annotation} slot is replaced by the matrix \code{annotations}.
}
\description{
\code{setAnnotations} changes the \code{annotation} slot from a \code{\link[=protdata-class]{protdata}} object to a prespecified matrix.
}

