% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{methods}
\name{setAnnotations,protLM-method}
\alias{setAnnotations,protLM-method}
\title{Change the annotation slot of a protLM object}
\usage{
\S4method{setAnnotations}{protLM}(object, annotations)
}
\arguments{
\item{object}{A \code{\link[=protLM-class]{protLM}} object of which the \code{annotation} slot needs to be modified.}

\item{annotations}{A matrix which will be inserted in the \code{annotation} slot of the \code{\link[=protLM-class]{protLM}} object. Note that the number of rows of this matrix should be equal to the number of elements in both the \code{accession} and the \code{model} slot.}
}
\value{
A \code{\link[=protLM-class]{protLM}} object of which the \code{annotation} slot is replaced by the matrix \code{annotations}.
}
\description{
\code{setAnnotations} changes the \code{annotation} slot from a \code{\link[=protLM-class]{protLM}} object to a prespecified matrix.
}

