% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{setData,protdata-method}
\alias{setData,protdata-method}
\title{Change the data slot of a protdata object}
\usage{
\S4method{setData}{protdata}(object, data)
}
\arguments{
\item{object}{A \code{\link[=protdata-class]{protdata}} object of which the \code{data} slot needs to be modified.}

\item{data}{A list of data frames which will be inserted in the \code{data} slot of the \code{\link[=protdata-class]{protdata}} object. Note that the length of this list should be equal to the number of elements in the \code{accession} slot and the number of rows in the \code{annotation} slot.}
}
\value{
A \code{\link[=protdata-class]{protdata}} object of which the \code{data} slot is replaced by the list of data frames \code{data}.
}
\description{
\code{setData} changes the \code{data} slot from a \code{\link[=protdata-class]{protdata}} object to a prespecified list of data frames.
}
