% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protLM.R
\docType{methods}
\name{setModels,protLM-method}
\alias{setModels,protLM-method}
\title{Change the model slot of a protLM object}
\usage{
\S4method{setModels}{protLM}(object, model)
}
\arguments{
\item{object}{A \code{\link[=protLM-class]{protLM}} object of which the \code{model} slot needs to be modified.}

\item{model}{A list of models which will be inserted in the \code{model} slot of the \code{\link[=protLM-class]{protLM}} object. Note that the length of this list should be equal to the number of elements in the \code{accession} slot and the number of rows in the \code{annotation} slot.}
}
\value{
A \code{\link[=protLM-class]{protLM}} object of which the \code{model} slot is replaced by the list of models \code{models}.
}
\description{
\code{setModels} changes the \code{model} slot from a \code{\link[=protLM-class]{protLM}} object to a prespecified list of models.
}
