% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protdata.R
\docType{methods}
\name{[,protdata,ANY,ANY,ANY-method}
\alias{[,protdata,ANY,ANY,ANY-method}
\title{Select elements from a protdata object}
\usage{
\S4method{[}{protdata,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{i}{Either a character or a factor vector containing accessions present in the \code{accessions} slot of the \code{\link[=protdata-class]{protdata}} object or a numeric vector containing indices pointing to certain elements in the the \code{\link[=protdata-class]{protdata}} object or a combination of both (deprecated).}

\item{object}{A \code{\link[=protdata-class]{protdata}} object of which certain elements will be selected.}
}
\value{
A reduced \code{\link[=protdata-class]{protdata}} object containing only the elements specified in \code{i}.
}
\description{
Select certain elements from a \code{\link[=protdata-class]{protdata}} object based on one or more accession and/or numeric indices.
}

