% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testANOVA.R
\name{test.ANOVA}
\alias{test.ANOVA}
\title{ANOVA for multiple contrasts}
\usage{
test.ANOVA(beta, vcov, df, sigma, L, anova.na.ignore = TRUE)
}
\arguments{
\item{beta}{A matrix with a single column containing parameter estimates.}

\item{vcov}{A matrix of the estimated covariances between the parameter estimates.}

\item{df}{A numeric value indicating the residual degrees of freedom.}

\item{sigma}{A numeric value indicating a residual standard deviation.}

\item{L}{A contrast matrix with the parameter levels as rows and a column for each contrast.}

\item{anova.na.ignore}{A logical indicating whether contrasts that cannot be fitted due to missing observations should be ignored when calculating an F value (\code{anova.na.ignore = TRUE}) or NA should be returned when at least one contrast cannot be estimated (\code{anova.na.ignore = FALSE}). Defaults to \code{TRUE}.}
}
\value{
A matrix with a single row and with three columns: (1) \code{AveExpr}, containing the average estimate of the contrasts; (2) \code{Fval}, containing the F-value corresponding to the contrasts; (3) \code{pval}, containing the p-value corresponding to the contrasts.
}
\description{
This function can test an ANOVA for multiple contrasts based on parameter estimates \code{beta}, a variance covariance matrix \code{vcov}, a number of degrees of freedom \code{df}, a residual error term \code{sigma} and a contrast matrix \code{L}.
}
\examples{
....
}
\references{
ANOVA
}
