% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testContrast.R
\name{test.contrast}
\alias{test.contrast}
\title{Test a contrast}
\usage{
test.contrast(beta, vcov, df, sigma, L, lfc = 0)
}
\arguments{
\item{beta}{A matrix with a single column containing parameter estimates.}

\item{vcov}{A matrix of the estimated covariances between the parameter estimates.}

\item{df}{A numeric value indicating the residual degrees of freedom.}

\item{sigma}{A numeric value indicating a residual standard deviation.}

\item{L}{A contrast matrix with the parameter levels as rows and a column for each contrast.}

\item{lfc}{The minimum log2-fold-change that is considered scientifically meaningful. Defaults to \code{0}.}
}
\value{
A matrix with a row for each contrast in \code{L} and with four columns: (1) \code{estimate}, containing the size estimates of the contrasts; (2) \code{se}, containing the estimated standard error on the contrast; (3) \code{Tval}, containing the T-value corresponding to the contrast; (4) \code{pval}, containing the p-value corresponding to the contrast.
}
\description{
This function can test a contrast based on parameter estimates \code{beta}, a variance covariance matrix \code{vcov}, a number of degrees of freedom \code{df}, a residual error term \code{sigma} and a contrast matrix \code{L}.
}
\examples{
#Load the protLM object protmodel:
protmodel <- data(modelRR, package="MSqRob")
betaVcovDfList <- getbetaVcovDfList(protmodel)
contrasts <- test.contrast(betaVcovDfList)
contrasts
}
\references{
David Ruppert, M.P. Want and R.J. Carroll.
Semiparametric Regression.
 Cambridge University Press, 2003.
}
