% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_contrast_adjust.R
\name{test.contrast_adjust}
\alias{test.contrast_adjust}
\title{Test a contrast and perform FDR correction}
\usage{
test.contrast_adjust(protLM, L, level = 0.05, method = "fdr",
  add.annotations = TRUE, custom_dfs = NULL, simplify = TRUE, lfc = 0,
  anova = FALSE, anova.na.ignore = TRUE, exp_unit = NULL,
  pars_df = NULL, satterthwaite = FALSE, lmerModFun = NULL,
  gradMethod = "simple", printProgress = FALSE, shiny = FALSE,
  message_extract = NULL, message_test = NULL)
}
\arguments{
\item{protLM}{An object of class \code{\link[=protLM-class]{protLM}}.}

\item{L}{A contrast matrix with the parameter levels as rows and a column for each contrast.}

\item{level}{The significance level at which the q-value needs to be controlled. Defaults to 5\%.}

\item{method}{Correction method. Can be abbreviated. Defaults to "fdr". To get all available methods, type \code{p.adjust.methods}. For more information on these methods, see the \code{\link{p.adjust}} function.}

\item{add.annotations}{A logical indicating whether the \code{annotations} slot of the \code{\link[=protLM-class]{protLM}} object should be added as extra columns to each matrix in the returned list of matrices. Defaults to \code{TRUE}.}

\item{custom_dfs}{An optional vector of length equal to the number of models in \code{protLM} containing the degrees of freedom that should be used for each model in the \code{protLM} object. By default (\code{custom_dfs=NULL}), \code{MSqRob} uses the trace of the Hat matrix to determine the degrees of freedom or the Satterhtwiate approximation (if \code{satterthwaite=TRUE}). \code{custom_dfs} allows you to overrule this and use your own custom approximation for the degrees of freedom.}

\item{simplify}{A logical indicating wheter, if there is only one contrast, a matrix should be returned instead of a list containing one matrix. Defaults to \code{TRUE}.}

\item{lfc}{The minimum (log2) fold-change that is considered scientifically meaningful. Defaults to \code{0}. Ignored when \code{anova = TRUE}.}

\item{anova}{A logical indicating whether the contrasts should be tested simultaneously in one F-test (\code{anova = TRUE}) or as separate t-tests (\code{anova = FALSE}). Defaults to \code{FALSE}.}

\item{anova.na.ignore}{A logical indicating whether contrasts that cannot be fitted due to missing observations should be ignored when calculating an F value (\code{anova.na.ignore = TRUE}) or NA should be returned when at least one contrast cannot be estimated (\code{anova.na.ignore = FALSE}). Defaults to \code{TRUE}. Ignored when \code{anova = FALSE}.}

\item{exp_unit}{The effect that in all models corresponds to the experimental unit. Only needed when one would like to calculate a more conservative way of estimating the degrees of freedom.
The default way of estimating the degrees of freedom (\code{exp_unit=NULL}) subtracts the total number of observations by the trace of the Hat matrix. However, often, observations are not completely independent. A more conservative way (\code{df_exp}) is defining on which level the treatments were executed and substracting all degrees of freedom lost due to between-treatement effects (\code{pars_df}) from the number of treatments.}

\item{pars_df}{Only used if exp_unit is not \code{NULL}. Character vector indicating all parameters in the models that are between-treatment effects in order to calculate a more conservative degrees of freedom (\code{df_exp}). If left to default (\code{NULL}), all parameters in the models will be asumed to be between-treatment effects (this is not adviced as the result will mostly be too conservative).}

\item{satterthwaite}{A logical indicating whether the Satterthwaite approximation for the degrees of freedom should be used instead of the classical trace of the Hat matrix. Defaults to \code{FALSE}.}

\item{lmerModFun}{Only used when \code{satterthwaite=TRUE}. \code{lmerModFun} indicates which deviance function should be used when calculating the Satterthwaite approximation for the degrees of freedom. The default (\code{NULL}) uses the lme4 \code{\link[=lme4::mkLmerDevfun]{mkLmerDevfun}} function to generate the deviance function. This parameter should only rarely, if ever, be changed.}

\item{gradMethod}{Only used when \code{satterthwaite=TRUE}. One of "Richardson", "simple", or "complex" indicating the method to use for the gradient calculation by numerical approximation during the calculation of the Satterthwaite approximation for the degrees of freedom. Defaults to "simple".}

\item{printProgress}{A logical indicating whether the R should print a message before calculating the contrasts for each accession. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message_extract}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the extraction of beta, vcov, df and sigma, or \code{NULL} to hide the current message (if any).}

\item{message_test}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the testing of the contrasts, or \code{NULL} to hide the current message (if any).}
}
\value{
A list of matrices, with each matrix in the list corresponding to a contrast in L. Each row of the matrix corresponds to an accession in the \code{\link[=protLM-class]{protLM}} object.
The \code{estimate} column contains the size estimate of the contrast, the \code{se} column contains the estimated standard error on the contrast, the \code{Tval} column contains the T-value corresponding to the contrast, the \code{pval} column holds the p-value corresponding to the contrast and the \code{qval} column holds the corrected p-values.
Each matrix is sorted from smalles to largest \code{pval} with \code{NA} values at the bottom of the matrices.
If \code{simplify=TRUE} and the \code{\link[=protLM-class]{protLM}} object contains only one element, the matrix is not present in a list.
}
\description{
A shortcut version for contrast testing and multiple testing correction using the \code{\link{test.protLMcontrast}}, \code{\link{prot.p.adjust}} and \code{\link{prot.signif}} functions sequentially.
}
