% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_contrast_stagewise.R
\name{test.contrast_stagewise}
\alias{test.contrast_stagewise}
\title{Perform stage-wise contrast testing}
\usage{
test.contrast_stagewise(protLM, L, add.annotations = TRUE, level = 0.05,
  method_stage1 = "fdr", squeezeVar = TRUE, par_squeeze = NULL,
  min_df = 1, custom_dfs = NULL, robust_var = TRUE, simplify = TRUE,
  lfc = 0, exp_unit = NULL, pars_df = NULL, satterthwaite = FALSE,
  printProgress = FALSE, shiny = FALSE, message_thetasS1 = NULL,
  message_squeezeS1 = NULL, message_updateS1 = NULL,
  message_extractS1 = NULL, message_testS1 = NULL,
  message_thetasS2 = NULL, message_squeezeS2 = NULL,
  message_updateS2 = NULL, message_extractS2 = NULL,
  message_testS2 = NULL, ...)
}
\arguments{
\item{protLM}{An object of class \code{\link[=protLM-class]{protLM}}.}

\item{L}{A contrast matrix with the parameter levels as rows and a column for each contrast.}

\item{add.annotations}{A logical indicating whether the \code{annotations} slot of the \code{\link[=protLM-class]{protLM}} object should be added as extra columns to each matrix in the returned list of matrices. Defaults to \code{TRUE}.}

\item{level}{The significance level at which the q-value needs to be controlled. Defaults to 5\%.}

\item{method_stage1}{Correction method to be used in the first stage ANOVA. Can be abbreviated. Defaults to "fdr". To get all available methods, type \code{p.adjust.methods}. For more information on these methods, see the \code{\link{p.adjust}} function.}

\item{squeezeVar}{A logical indicating whether residual standard deviations should be squeezed together by computing empirical Bayes posterior means. For more information, see limma's \code{\link{squeezeVar}} function. Defaults to \code{TRUE}.}

\item{par_squeeze}{Character vector indicating which model parameters need to be squeezed. When squeezing random effects, provide their names. Fixed effects are present in shrinkage groups, e.g. ridgeGroup.1. If you want them to be squeezed as well, provide the names of the shrinkage groups that need to be squeezed. The default \code{NULL} indicates that no parameters will be squeezed.}

\item{min_df}{A numeric value indicating the minimal number of residual degrees of freedom that is required for a protein to be taken into account for the squeezing of the variances. Only used when \code{squeezeVar} is set to \code{TRUE}. Defaults to \code{1}.}

\item{robust_var}{A logical indicating wheter the squeezing of the variances should be robust to the presence of outlier sample variances. Defaults to \code{TRUE}.}

\item{simplify}{A logical indicating wheter, if there is only one contrast, a matrix should be returned instead of a list containing one matrix. Defaults to \code{TRUE}.}

\item{lfc}{The minimum (log2) fold-change that is considered scientifically meaningful. Defaults to \code{0}. Ignored when \code{anova = TRUE}.}

\item{exp_unit}{The effect that in all models corresponds to the experimental unit. Only needed when one would like to calculate a more conservative way of estimating the degrees of freedom.
The default way of estimating the degrees of freedom (\code{exp_unit=NULL}) subtracts the total number of observations by the trace of the Hat matrix. However, often, observations are not completely independent. A more conservative way (\code{df_exp}) is defining on which level the treatments were executed and substracting all degrees of freedom lost due to between-treatement effects (\code{pars_df}) from the number of treatments.}

\item{pars_df}{Only used if exp_unit is not \code{NULL}. Character vector indicating all parameters in the models that are between-treatment effects in order to calculate a more conservative degrees of freedom (\code{df_exp}). If left to default (\code{NULL}), all parameters in the models will be asumed to be between-treatment effects (this is not adviced as the result will mostly be too conservative).}

\item{satterthwaite}{A logical indicating whether the Satterthwaite approximation for the degrees of freedom should be used instead of the classical trace of the Hat matrix. Defaults to \code{FALSE}.}

\item{printProgress}{A logical indicating whether the R should print a message before calculating the contrasts for each accession. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message_thetasS1}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the extraction of the variances during stage 1, or \code{NULL} to hide the current message (if any).}

\item{message_squeezeS1}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the squeezing of the variances during stage 1, or \code{NULL} to hide the current message (if any).}

\item{message_updateS1}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the updating of the models during stage 1, or \code{NULL} to hide the current message (if any).}

\item{message_extractS1}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the extraction of beta, vcov, df and sigma during stage 1, or \code{NULL} to hide the current message (if any).}

\item{message_testS1}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the testing of the contrasts during stage 1, or \code{NULL} to hide the current message (if any).}

\item{message_thetasS2}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the extraction of the variances during stage 2, or \code{NULL} to hide the current message (if any).}

\item{message_squeezeS2}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the squeezing of the variances during stage 2, or \code{NULL} to hide the current message (if any).}

\item{message_updateS2}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the updating of the models during stage 2, or \code{NULL} to hide the current message (if any).}

\item{message_extractS2}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the extraction of beta during stage 2, vcov, df and sigma, or \code{NULL} to hide the current message (if any).}

\item{message_testS2}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the testing of the contrasts during stage 2, or \code{NULL} to hide the current message (if any).}

\item{...}{Additional arguments to be passed to the squeezeVarRob function internally.}
}
\value{
A list of matrices, with each matrix in the list corresponding to a contrast in L. Each row of the matrix corresponds to a protein in the \code{\link[=protLM-class]{protLM}} object.
The \code{estimate} column contains the size estimate of the contrast, the \code{se} column contains the estimated standard error on the contrast, the \code{Tval} column contains the T-value corresponding to the contrast, the \code{pval} column holds the p-value corresponding to the contrast and the \code{qval} column holds the corrected p-values.
Each matrix is sorted from smalles to largest \code{pval} with \code{NA} values at the bottom of the matrices.
If \code{simplify=TRUE} and the \code{\link[=protLM-class]{protLM}} object contains only one element, the matrix is not present in a list.
}
\description{
A shortcut version for contrast testing and multiple testing correction using first \code{\link{test.contrast_adjust}} to perform an ANOVA as a first stage.
In the second stage, the functions \code{\link{test.protLMcontrast}}, \code{\link{prot.p.adjust_protwise}} and \code{\link{prot.signif}} are used sequentially only on the proteins retained in the first stage.
This approach is more powerful than the classic \code{\link{test.contrast_adjust}} when testing for contrasts where related proteins are expected to be differentially abundant. This approach is also better for identifying interactions.
}

