% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeezePars.R
\name{update.lmerMod}
\alias{update.lmerMod}
\title{Update lmerMod object with new thetas without changing the pointers of other objects}
\usage{
\method{update}{lmerMod}(object, theta, ...)
}
\arguments{
\item{object}{A \code{\link[=lmerMod-class]{lmerMod}} object of which residual standard deviations and/or model parameters need to be squeezed.}

\item{theta}{A numeric vector of theta values (parameter variances) that will replace the theta values in the \code{\link[=lmerMod-class]{lmerMod}} object.}

\item{...}{Other arguments to be passed to the \code{\link{update.default}} function if the \code{theta} input argument would be missing.}
}
\value{
A \code{\link[=lmerMod-class]{lmerMod}} object of which the parameter variances are replaced by the ones given in the \code{theta} input argument.
}
\description{
This function changes the \code{theta} parameter of a \code{\link[=lmerMod-class]{lmerMod}} object to a given vector and updates the model accordingly.
}
\examples{
....
}
\references{
Bolker (2016). Wald errors of variances, https://rpubs.com/bbolker/waldvar
}

