% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSnSet2protdata.R
\name{MSnSet2protdata}
\alias{MSnSet2protdata}
\title{MSnSet to protein data}
\usage{
MSnSet2protdata(MSnSet, accession, annotations = NULL)
}
\arguments{
\item{MSnSet}{An \code{\link[=MSnSet-class]{MSnSet}} object that needs to be converted to a \code{\link[=protdata-class]{protdata}} object.}

\item{accession}{A character string or numeric index indicating the column in the fData slot of object  \code{\link[=MSnSet-class]{MSnSet}} that contains the identifiers by which the data should be grouped in the resulting \code{\link[=protdata-class]{protdata}} object (typically the protein identifier).}

\item{annotations}{A vector of character strings or numeric indices indicating the columns in the fData slot of object  \code{\link[=MSnSet-class]{MSnSet}} that contain additional information on the accessions (typically protein names, gene names, gene ontologies,...) that should be retained. In case multiple values of the same annotation column would exist for a unique accession, these values are pasted together. Defaults to \code{NULL}, in which case no annotations will be added.}
}
\value{
A \code{\link[=protdata-class]{protdata}} object.
}
\description{
Converts an \code{\link[=MSnSet-class]{MSnSet}} object to a a \code{\link[=protdata-class]{protdata}} object, which can be used for further analysis.
}
\examples{
#This example will convert MSnSet object peptides into a protdata object proteins.
library(MSnbase)
colInt <- grepEcols(system.file("extdata/CPTAC", "peptides.txt", package = "MSqRob"), pattern="Intensity.", split = "\\t")
#Import the data as an MSnSet object
peptides <- readMSnSet2(system.file("extdata/CPTAC", "peptides.txt", package = "MSqRob"), ecol = colInt, sep = "\\t")
fData(peptides) <- fData(peptides)[,c("Proteins","Sequence","PEP")]
#Do this only for the first 50 peptide to save execution time
proteins <- MSnSet2protdata(peptides[1:50], "Proteins")
}

