% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_MaxQuant.R
\name{read2MSnSet}
\alias{read2MSnSet}
\title{Import a file and convert it to an MSnSet object}
\usage{
read2MSnSet(file, pattern, remove_pattern = FALSE, sep = "\\t",
  shiny = FALSE, message = NULL)
}
\arguments{
\item{file}{The name of a file. For more details about how this argument can be specified, see \code{\link[utils]{read.table}}.}

\item{pattern}{A character string containing a regular expression that will be matched to the file's header. The columns matching the expression should be the columns containing the peptide intensity values.}

\item{remove_pattern}{A logical indicating whether the expression in "pattern" should be removed from the column names in the resulting \code{\link[=MSnSet-class]{MSnSet}} object. Defaults to \code{FALSE}.}

\item{sep}{A character indicating the field separator character. Values on each line of the file are separated by this character.The  Defaults to "\\t", indicating a tab-delimited input file.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user, or \code{NULL} to hide the current message (if any).}
}
\value{
An object of class \code{\link[=MSnSet-class]{MSnSet}}.
}
\description{
Imports a file and converts it into an \code{\link[=MSnSet-class]{MSnSet}} object (Gatto et al., 2012).
}
\references{
Gatto L, Lilley KS. MSnbase - an R/Bioconductor package for isobaric tagged mass spectrometry data visualization, processing and quantitation. Bioinformatics. 2012 Jan 15;28(2):288-9. \url{https://doi.org/10.1093/bioinformatics/btr645}.
}
