% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeezePars.R
\name{squeezePars}
\alias{squeezePars}
\title{Squeeze variances and other parameters}
\usage{
squeezePars(protLM, par_squeeze = NULL, squeezeVar = TRUE, min_df = 1,
  robust_var = TRUE, printProgress = FALSE, shiny = FALSE,
  message_thetas = NULL, message_squeeze = NULL, message_update = NULL,
  ...)
}
\arguments{
\item{protLM}{A \code{\link[=protLM-class]{protLM}} object of which residual standard deviations and/or model parameters need to be squeezed.}

\item{par_squeeze}{Character vector indicating which model parameters need to be squeezed. When squeezing random effects, provide their names. Fixed effects are present in shrinkage groups, e.g. ridgeGroup.1. If you want them to be squeezed as well, provide the names of the shrinkage groups that need to be squeezed. The default \code{NULL} indicates that no parameters will be squeezed.}

\item{squeezeVar}{A logical indicating whether the residual standard deviation of all models should be squeezed towards a common value. Defaults to \code{TRUE}. If set to \code{FALSE}, residual standard deviations will not be squeezed.}

\item{min_df}{A numeric value indicating the minimal degrees of freedom that will be taken into account for calculating the prior degrees of freedom and prior variance.}

\item{robust_var}{A logical indicating wheter the estimation of the prior degrees of freedom and the prior variance (needed for shrinkage) should be robustified against outlier variances. Defaults to \code{TRUE}.}

\item{printProgress}{A logical indicating whether the R should print a message before performing each preprocessing step. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message_thetas}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the extraction of the variances, or \code{NULL} to hide the current message (if any).}

\item{message_squeeze}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the squeezing of the variances, or \code{NULL} to hide the current message (if any).}

\item{message_update}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the updating of the models, or \code{NULL} to hide the current message (if any).}

\item{...}{Other arguments to be passed to the \code{update_protLM} function internally.}
}
\value{
An updated protLM object with squeezed variances and/or squeezed parameters (or the input \code{\link[=protLM-class]{protLM}} object if \code{par_squeeze=NULL} and \code{squeezeVar=FALSE}).
}
\description{
Squeeze the standard deviations of the models towards a common value using empirical Bayes moderation similar to the \code{\link[=limma]{limma}} package.
Information is borrowed over all accessions in the \code{\link[=protLM-class]{protLM}} object in order to stabilize variances. These stabilized variances are called posterior variances. Options are provided to also perform the same squeezing over the random effects and for the penalty of the shrunken fixed effects.
}
\examples{
....
}
\references{
....
}

