% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeezePars.R
\name{squeezeThetas}
\alias{squeezeThetas}
\title{Squeeze variances and other parameters}
\usage{
squeezeThetas(thetas, df_thetas, vars, df_vars, squeezeVar = TRUE,
  min_df = 1, robust_var = TRUE, printProgress = FALSE, shiny = FALSE,
  message = NULL, ...)
}
\arguments{
\item{thetas}{A numeric matrix wherein each column contains variances for a different parameter.}

\item{df_thetas}{A numeric matrix wherein each column contains degrees of freedom corresponding to the \code{thetas}.}

\item{vars}{A numeric vector containing residual variances.}

\item{df_vars}{A numeric vector containing residual degrees of freedom.}

\item{squeezeVar}{A logical indicating whether the residual variances as given by the argument \code{vars} should be squeezed towards a common value. Defaults to \code{TRUE}. If set to \code{FALSE}, \code{vars} will not be squeezed and the \code{vars_post} slot will contain the same values as the \code{vars} input argument.}

\item{min_df}{A numeric value indicating the minimal degrees of freedom that will be taken into account for calculating the prior degrees of freedom and prior variance.}

\item{robust_var}{A logical indicating wheter the estimation of the prior degrees of freedom and the prior variance (needed for shrinkage) should be robustified against outlier variances. Defaults to \code{TRUE}.}

\item{printProgress}{A logical indicating whether the R should print a message before performing each preprocessing step. Defaults to \code{FALSE}.}

\item{shiny}{A logical indicating whether this function is being used by a Shiny app. Setting this to \code{TRUE} only works when using this function in a Shiny app and allows for dynamic progress bars. Defaults to \code{FALSE}.}

\item{message}{Only used when \code{printProgress=TRUE} and \code{shiny=TRUE}. A single-element character vector: the message to be displayed to the user during the squeezing of the variances and degrees of freedom, or \code{NULL} to hide the current message (if any).}
}
\value{
A named list with 4 slots. The first slot \code{thetas_post} contains a matrix with in each column the squeezed variances corresponding to the \code{thetas} input argument.
The second slot \code{df_thetas_post} contains a matrix of similar structure to \code{thetas_post} but containing the posterior degrees of freedom corresponding to the \code{df_thetas} input argument. If \code{squeezeVar=TRUE}, the third slot \code{vars_post} contains a vector of posterior residual variances for each accession. If \code{squeezeVar=FALSE}, it contains the residual variances given in the \code{vars} input argument.
If \code{squeezeVar=TRUE}, the fourth slot \code{df_vars_post} contains a vector of posterior residual degrees of freedom. If \code{squeezeVar=FALSE}, it contains the residual degrees of freedom given in the \code{df_vars} input argument.
}
\description{
Squeeze a set of sample variances towards a common value using empirical Bayes moderation similar to the \code{\link[=limma]{limma}} package.
This can be done for residual variances, random effects or shrunken fixed effects. The resulting variances are called posterior variances.
}
\examples{
....
}
\references{
....
}

